cleaned up measurement types
